+++ packages/terra-application-docs/src/terra-dev-site/test/terra-dev-site/navigationPrompt.test.jsx
@@ -1,5 +1,5 @@
 import React, { useState } from 'react';
-import NavigationPrompt from 'terra-application/lib/navigation-prompt';
+import NavigationPrompt from 'terra-application/navigation-prompt';
 
 const ExampleComponent = () => {
   const [hasPendingState, setHasPendingState] = useState(false);