+++ packages/terra-application/src/application-base/private/intlLoaders.js
@@ -1,6 +1,6 @@
 import hasIntlData from 'intl-locales-supported';
 import loadLocaleData from './loadLocaleData';
-import logger from '../../utils/logger';
+import { Logger } from '../../utils';
 
 const supportedIntlConstructors = (polyfill) => {
   /**
@@ -52,7 +52,7 @@ const loadIntl = (locale, polyfill) => {
   if (!hasIntlData([locale], supportedIntlConstructors(polyfill))) {
     return loadLocaleData(locale, polyfill).catch((error) => {
       if (fallbackLocale) {
-        logger.warn(`${error.message} Using ${fallbackLocale} data as the fallback locale data.`);
+        Logger.warn(`${error.message} Using ${fallbackLocale} data as the fallback locale data.`);
         if (!hasIntlData([fallbackLocale], supportedIntlConstructors(polyfill))) {
           return loadLocaleData(fallbackLocale, polyfill);
         }
@@ -62,7 +62,7 @@ const loadIntl = (locale, polyfill) => {
 
       return Promise.resolve();
     }).catch((error) => {
-      logger.warn(`${error.message} Using en data as the fallback locale data.`);
+      Logger.warn(`${error.message} Using en data as the fallback locale data.`);
 
       if (!hasIntlData(['en'], supportedIntlConstructors(polyfill))) {
         return loadLocaleData('en', polyfill);