+++ packages/terra-application/src/terra-dev-site/test/application-base/ApplicationBaseTest.test.jsx
@@ -1,11 +1,11 @@
 import React, { useState, useContext } from 'react';
-import { ActiveBreakpointContext } from 'terra-application/lib/breakpoints';
-import ApplicationLoadingOverlay from 'terra-application/lib/application-loading-overlay';
-import ApplicationStatusOverlay from 'terra-application/lib/application-status-overlay';
-import ApplicationBase from 'terra-application/lib/application-base';
-import NavigationPrompt from 'terra-application/lib/navigation-prompt';
+import { ActiveBreakpointContext } from 'terra-application/breakpoints';
+import ApplicationLoadingOverlay from 'terra-application/application-loading-overlay';
+import ApplicationStatusOverlay from 'terra-application/application-status-overlay';
+import ApplicationBase from 'terra-application/application-base';
+import NavigationPrompt from 'terra-application/navigation-prompt';
 import { useIntl } from 'react-intl';
-import { ThemeContext } from 'terra-application/lib/theme';
+import { ThemeContext } from 'terra-application/theme';
 
 const ApplicationContentTest = () => {
   const [isLoading, setIsLoading] = useState(false);