+++ packages/terra-application/tests/jest/utils/logger/logger.test.js
@@ -1,4 +1,4 @@
-import Logger, { initializeLogger } from '../../../../src/utils/logger';
+import { Logger, initializeLogger } from '../../../../src/utils';
 
 /* eslint-disable no-console */
 describe('Logger', () => {