+++ packages/terra-dev-site/src/webpack/plugin/SitePlugin.js
@@ -28,12 +28,10 @@ const urlQueue = [];
  * Updates the webpack options with defaults that terra-dev-site requires.
  */
 class SitePlugin {
-  constructor({
-    entry,
-    config,
-  }) {
+  constructor({ entry, config }) {
     // Apply defaults to the config.
     this.siteConfig = config;
+
     const { pathPrefix, titleConfig } = this.siteConfig;
     this.entry = entry;
 
@@ -67,6 +65,7 @@ class SitePlugin {
     distributionFolder,
     basename,
     isWebpack5,
+    useDefaultWebpackResolver,
   }) {
     if (oneTimeSetupComplete) {
       return;
@@ -96,10 +95,7 @@ class SitePlugin {
           // Only the first loader will apply and no others.
           oneOf: [{
             test: /\.mdx$/,
-            use: [
-              babelLoader,
-              mdxLoader,
-            ],
+            use: [babelLoader, mdxLoader],
           }, {
             test: /\.md$/,
             oneOf: [
@@ -134,19 +130,13 @@ class SitePlugin {
             ],
           }, {
             resourceQuery: '?dev-site-example',
-            use: [
-              babelLoader,
-              'devSiteExample',
-            ],
+            use: [babelLoader, 'devSiteExample'],
           }, {
             test: /\.json$/,
             // this bypasses the default json loader
             type: 'javascript/auto',
             resourceQuery: '?dev-site-package',
-            use: [
-              babelLoader,
-              'devSitePackage',
-            ],
+            use: [babelLoader, 'devSitePackage'],
           }, {
             resourceQuery: '?dev-site-props-table',
             use: [
@@ -187,10 +177,8 @@ class SitePlugin {
               }),
             ]
             : [],
-          // Alias the local package to allow imports to reference the file as if it was imported from node modules.
-          new LocalPackageAliasPlugin({
-            rootDirectories,
-          }),
+          // // Alias the local package to allow imports to reference the file as if it was imported from node modules.
+          ...(useDefaultWebpackResolver ? [] : [new LocalPackageAliasPlugin({ rootDirectories })]),
         ],
       },
       // add the path to search for dev site loaders
@@ -269,7 +257,7 @@ class SitePlugin {
     // Strip the trailing / from the public path.
     let basename = publicPath.slice(0, -1);
 
-    const { sourceFolder, distributionFolder } = this.siteConfig;
+    const { sourceFolder, distributionFolder, useDefaultWebpackResolver } = this.siteConfig;
 
     // Since there can be multiple dev site plugins this config we only want to do once for all of them.
     SitePlugin.applyOneTimeSetup({
@@ -278,6 +266,7 @@ class SitePlugin {
       distributionFolder,
       basename,
       isWebpack5,
+      useDefaultWebpackResolver,
     });
 
     // Get the list of apps excluding this current app.