+++ packages/terra-application-docs/CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * Changed
   * Set minimum Node version to 14.
+  * Updated docs to utilize the new terra-application imports.
 
 ## 3.1.0 - (June 22, 2022)