+++ packages/terra-application/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Unreleased
 
+* Breaking Changes:
+  * Added subpath exports for terra-application submodules. This replaces the old exports that utilize the exact path.
+  * Removed default `ApplicationBase` import. `ApplicationBase` will now need to be imported from `terra-application/application-base`.
+
 * Changed
   * Dropped support for Node 8, 10 & 12. Node 14 is now the minimum required version.
   * Locked `intl-locales-supported` to `1.8.11`.